+++ R/nhanesdata-package.R
@@ -1,0 +1,14 @@
+#' @keywords internal
+"_PACKAGE"
+
+## usethis namespace: start
+#' @importFrom arrow read_parquet
+#' @import dplyr
+#' @importFrom nhanesA nhanes nhanesSearch nhanesSearchVarName
+#' @importFrom rlang abort
+#' @importFrom scales comma
+#' @importFrom stringr str_to_upper
+#' @importFrom srvyr as_survey_design
+#' @importFrom tibble tibble
+## usethis namespace: end
+NULL